Rename GlobalDrag to GlobalDragComponent
This keeps the naming similar to KeyComponent, ResizeComponent,
ScrollComponent...
