+++ src/components/viewport/GlobalDragComponent.js
